Solving int/long definition

+++ src/hm.h
@@ -26,12 +26,12 @@ double* beta0;
 double* gamma0;
 
 typedef struct{
-	int tid;  // thread id
-        int nfeaturesperthread;  // number of peaks processed [tid*npeaks ... (tid+1)*npeaks-1]
-	int nfeatures;  // total peaks
-	int* cumloci;
-	int Pi;
-	int Pj;
+	long tid;  // thread id
+        long nfeaturesperthread;  // number of peaks processed [tid*npeaks ... (tid+1)*npeaks-1]
+	long nfeatures;  // total peaks
+	long* cumloci;
+	long Pi;
+	long Pj;
 	double* X0;
 	double* BF;
